[#107][Refactoring] - modifying an existing non-course event into a course event now correctly updates the user's course list
